Количество сделано с плавающей точкой, взамен целого
Тип "Количество" в списках услуг сделан вещественным, чтобы можно было
указывать не дробные числа, например: 12,5 м2 или 0,5 месяца

+++ newaprashka/PayAccrual.cs
@@ -202,7 +202,7 @@ namespace bazar
 						paid = rdr.GetDecimal ("paid");
 					else
 						paid = 0;
-					accrual = rdr.GetInt32("count") * rdr.GetDecimal("price");
+                    accrual = rdr.GetDecimal("count") * rdr.GetDecimal("price");
 					sum = Convert.ToDouble (accrual - paid);
 					if(sum <= 0)
 						continue;
